chapters-writing-skills/skills/interactive-fiction-writer/pyproject.toml
[project]
name = "interactive-fiction-writer"
version = "0.1.0"
description = "Interactive fiction writer using LLM to generate dialogue and narration for story skeletons"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.84.0",
    "python-dotenv>=1.0.0",
    "json-repair>=0.39.1",
    "openai>=2.26.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
]
